v3.4.2: 构建系统优化 + CI/CD集成 + 性能基线
核心改进:
- CMake配置时间 25s  9.3s (-63%)
- mingw64 64位工具链自动检测
- OpenSSL 3.3.1查找修复
- 性能基线建立 (BLAKE2b +31.77%, SM3 +51.53%)
- GitHub Actions CI/CD + 性能门禁
- Per-target -Werror策略 (thirdparty豁免)
- 172个编译警告修复

性能数据 (10MB):
- SHA3-256: 493 MB/s (-8.29% vs OpenSSL, 优化中)
- SHA3-512: 292 MB/s (-1.16%, near-parity)
- BLAKE2b: 934 MB/s (+31.77%, best-in-class)
- SM3: 355 MB/s (+51.53%, outstanding)

新增文件:
- docs/PERFORMANCE_BASELINE.md
- .github/workflows/ci.yml
- scripts/check_performance.py
- docs/releases/v3.4.2-release.md

Hash测试: 29/29通过
下一步: SHA3-256优化目标 567 MB/s (v3.5.0)

+++ src/advanced/zk/ffs/ffsa.cpp
@@ -156,7 +156,7 @@ void ffsa_get_secrets(mpz_vec_t& s, int8_t k, const mpz_t n)
                 continue;
             
             // Check for duplicates
-            for(int8_t j = 0; j < i; j++)
+            for(size_t j = 0; j < static_cast<size_t>(i); j++)
             {
                 if(mpz_cmp(intermediate, *s[j]) == 0)
                 {
